Return file information about target not link
If a file object refers to a symbolic link, any methods which
make use of #stat (such as #size) will return information about
the link rather than the target file it refers to.

This leads to a situation where, for example, Rack complains
that the length of the HTTP body and the Content-Length header
don't match up.

In order to support symlinked files, it therefore makes more
sense to use 'stat' instead of 'lstat' here.