Revert back to using stat instead of lstat
Following conversations with Mark, in the short term it is
preferable to display the size, perms etc. of a symlink's
target rather than those of the symlink itself. However,
eventually we want to customise the finder UI to allow the
user to find the item to which a symlink refers.