Handle downloading symlinked files in Api class
Within an instance of Volume (which doesn't know anything about
any of the other Volumes) it's tricky to obtain information
about the target of a symlink when it's in a different Volume.
Handling this in the Api class avoids the problem and eliminates
the need to add special cases for symlinks to methods like
File#mimetype.