Store symlinks in zip archives as true symlinks
By default the UNIX 'zip' command dereferences symlinks, i.e.
stores the file or directory referred to by the link. This
creates the possibility of a rogue symlink to '/' (or similar)
causing the entire root filesystem to be archived.

By passing the '-y' flag to zip, symlinks are stored as actual
symlinks, comparably to the way tar does it. Unfortunately how
compatible this is is unclear; the man page for zip(1) only says
that this feature is supported on "UNIX and VMS (V8.3 and later)".